+++ r/NEWS.md
@@ -17,7 +17,7 @@
   under the License.
 -->
 
-# arrow 0.16.0.9000
+# arrow 0.17.0
 
 ## Feather v2